requirements.txt
solana==0.30.2
requests==2.31.0
python-dotenv==1.0.0
aiohttp==3.9.1
asyncio==3.4.3
web3==6.11.1
base58==2.1.1
typing-extensions==4.8.0
dataclasses==0.6
pydantic==2.5.2
tenacity==8.2.3
websockets>=9.0,<12.0 